arregle errores y habili funciones de admin

+++ app/src/main/java/com/example/apptiendadeportiva_grupo10/model/Producto.kt
@@ -4,6 +4,6 @@ data class Producto (
     val id: Int,
     val nombre: String,
     val descripcion: String?,
-    val precio: Double,
+    val precio: Int,
     val imagen: String?
 )
